o Address comments, make one test file for vector calculus
